requirements.txt
faiss-cpu
pypdf
langchain
groq